services/micro/auth/requirements.txt
flask==3.1.3
psycopg2-binary==2.9.9
qrcode==7.4.2
Pillow==10.4.0
